reorganised some functions
